Servers/server2/mqtt-to-influxdb/requirements.txt
paho-mqtt~=1.5.0
influxdb~=5.3.0